api/requirements.txt
fastapi
uvicorn
google-generativeai
supabase
python-dotenv
requests
pydantic
pydantic-settings
